refactor: keep original file name for apply diff tool
